+++ starkbank/event/__event.py
@@ -12,10 +12,11 @@ from ..boleto.log.__log import _resource as _boleto_log_resource
 from ..invoice.log.__log import _resource as _invoice_log_resource
 from ..deposit.log.__log import _resource as _deposit_log_resource
 from ..transfer.log.__log import _resource as _transfer_log_resource
-from ..brcodepayment.log.__log import _resource as _brcodepayment_payment_log_resource
+from ..taxpayment.log.__log import _resource as _tax_payment_log_resource
+from ..boletoholmes.log.__log import _resource as _boleto_holmes_log_resource
 from ..boletopayment.log.__log import _resource as _boleto_payment_log_resource
 from ..utilitypayment.log.__log import _resource as _utility_payment_log_resource
-from ..boletoholmes.log.__log import _resource as _boleto_holmes_log_resource
+from ..brcodepayment.log.__log import _resource as _brcodepayment_payment_log_resource
 from ..error import InvalidSignatureError
 from ..utils import cache
 
@@ -28,6 +29,7 @@ _resource_by_subscription = {
     "brcode-payment": _brcodepayment_payment_log_resource,
     "boleto-payment": _boleto_payment_log_resource,
     "utility-payment": _utility_payment_log_resource,
+    "tax-payment": _tax_payment_log_resource,
     "holmes": _boleto_holmes_log_resource,
 }
 
@@ -39,7 +41,7 @@ class Event(Resource):
     list all generated updates on entities.
     ## Attributes:
     - id [string]: unique id returned when the event is created. ex: "5656565656565656"
-    - log [Log]: a Log object from one the subscription services (TransferLog, InvoiceLog, DepositLog, BoletoLog, BoletoHolmesLog, BrcodePaymentLog, BoletoPaymentLog or UtilityPaymentLog)
+    - log [Log]: a Log object from one the subscription services (TransferLog, InvoiceLog, DepositLog, BoletoLog, BoletoHolmesLog, BrcodePaymentLog, BoletoPaymentLog, UtilityPaymentLog or TaxPaymentLog)
     - created [datetime.datetime]: creation datetime for the notification event. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     - is_delivered [bool]: true if the event has been successfully delivered to the user url. ex: False
     - subscription [string]: service that triggered this event. ex: "transfer", "utility-payment"