+++ tests/utils/taxPayment.py
@@ -1,10 +1,10 @@
 from .nonBoletoPayment import generateExampleNonBoletoPaymentsJson
 
 
-def generateExampleUtilityPaymentsJson(n=1, amount=None, next_day=False):
+def generateExampleTaxPaymentsJson(n=1, amount=None, next_day=False):
     return generateExampleNonBoletoPaymentsJson(
         n=n,
         amount=amount,
         next_day=next_day,
-        is_tax=False,
+        is_tax=True,
     )